feat: reward update logic

+++ app/src/main/java/com/mobdeve/s12/mp/gamification/localdb/TaskSkillReward.kt
@@ -21,7 +21,7 @@ import kotlinx.coroutines.flow.Flow
 data class RewardEntity (
     @ColumnInfo(name = "taskId") val taskId: Long, // Foreign key referencing the task
     @ColumnInfo(name = "skillId") val skillId: Long, // Foreign key referencing the skill
-    val reward: Float
+    val reward: Int
 )
 
 // Queries
