+++ app/src/main/java/com/mobdeve/s12/mp/gamification/model/Reward.kt
@@ -5,9 +5,9 @@ import java.io.Serializable
 data class Reward(
     var task: Task,
     var skill : Skill,
-    var xp: Float
+    var xp: Int
 ) : Serializable {}
 
 fun createDefaultSkillReward(t: Task, sk: Skill = Skill(0, "Rewarded skill", "Description", )) : Reward{
-    return Reward(t, sk, 10.0f)
+    return Reward(t, sk, 10)
 }
